move poorly named `app` to `musicfig`
