+++ musicfig/colors.py
