+++ musicfig/templates/favicon.png
