+++ run.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
 
 
-from app import app
+from musicfig import app
 app.run(host='0.0.0.0', port=5000, debug=True)
